fix(review): address standards and spec review findings for Phase 1 soundness
